ustctl: add base trigger support
The trigger feature allows the user to get notify through a pipe that an
callsite was reached. It's really similar to a regular UST event in that
firing can be controlled by filters and exclusions. It diverges because
firing a trigger does not end up writing to the tracing buffers. In
fact, triggers live outside of any sessions (i.e. no session is needed).

Definitions
===========
A `Trigger` is defined as an event description (event name, log level,
etc.) and a unique trigger id.

A `Trigger group` is a set of triggers sharing the same pipe to the
liblttng-ust-ctl user.

Sequence
========
  trigger_group_handle = ustctl_create_trigger_group(notification_fd)
  trigger_handle = ustctl_create_trigger(trigger_group_handle, trigger_id);
  ustctl_set_filter(trigger_handle, filter)
  ustctl_set_exclusion(trigger_handle, exclusion)
  ustctl_enable(trigger_handle)
  ...
  ustctl_disable(trigger_handle)
  ustctl_release_object(trigger_handle)
  ustctl_release_object(trigger_group_handle)

WIP

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>
Change-Id: I493d784cc6830cd6c7979f6f08b209521676f05c