Rename `enum lttng_enable_type` to `_format_type`
Currently, the `_type` field is used to express in what format the event
rule was provided, either the full event name or a glob pattern.

Future commit will introduce the concept trigger enabler that will be
along side regular event enabler.
The use of the `_type` suffix might become confusing with this new class
of enablers.

So this commit rename the  `_type` field to `_format_type`.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>
Change-Id: I4d1e97da03df3983cba8b65c35d42818e0a967a5