Fix: only link triggers to enabler when enabled
Since we were syncing enablers on every actions on the trigger enablers,
we ended up with multiple version of the enabler link to the trigger
objects.

Here is the problematic scenario:
1. User creates trigger enabler (TE) on wildcard event (event_*)
2. Lib creates a trigger event_a, object and link to TE,
2. Lib creates a trigger event_b, object and link to TE,
3. User attaches an exclusion on TE to exclude event_a
4. User enables TE
5. Lib iterates over all triggers and checked for linked and enabled
   enabler and enable them.
6. User sees triggers for event_a AND event_b when only event_b is
   expected.

This problem is due to the fact the trigger objects are linked to
enablers before they are made unmutable by the enable() API call.

To fix that, only link triggers when enablers are enabled.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>
Change-Id: I5a87bda3eab2f4b893172d6ab482776f35839884