Remove event context dead code
The event context is never wired up by the lttng ust abi, so it is never
actually used. Remove this dead code which adds complexity for no
reason.

This code is all internal to lttng-ust (not exposed to probe providers),
which makes it easy to add this back if needed in the future.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: I52e93835ffe6d3040ead7fb8d3113ab20a555019