Custom upgrade: wire up configure --enable-custom-upgrade-conflicting-symbols
This custom configure switch allows selecting between those two
behaviors for symbols which clash between lttng-ust 2.12 and 2.13:

Disabled (default): only emit and use symbols with "1" suffix. This
is meant to be used alonside lttng-ust 2.12.

Enabled: emit both symbols with and without "1" suffix, and use the
symbol without suffix. This is meant to be used after there are no
more users of lttng-ust 2.12 to gradually transition back to an upstream
lttng-ust 2.13 as applications are rebuilt.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: I7ed02b2f0f9bd8c7b7b24fce28d2f23d61c970a3