Custom upgrade: chain load liblttng-ust-{fd,fork}.so.0 with dlopen
dlopen liblttng-ust.so.0 in the constructor of the interposer librairies
and explicitly call the symbols from this library when appropriate.

As a side-note, chain linking the interposer ld_preload librairies on
their previous SONAME0 versions doesn't work as expected because the
order of the symbol resolution with dlsym always favors the preloaded
library symbols. This is why explicit use of dlopen/dlsym is favored
instead.

Change-Id: Ie8802134c3037cfd431a949f469d3e64e958888b
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>