Fix: lazily dlopen 2.12 lttng-ust from liblttng-ust-fd.so symbol overrides
Fix warnings:

/usr/lib64/liblttng-ust-common.so.1: undefined symbol: lttng_ust_safe_close_fd

Lazily dlopen liblttng-ust.so.0 when close/fclose are used early before
execution of liblttng-ust-fd.so's constructor, and use it to perform the
perform the required dlsym() lookups to find the 2.12 LTTng-UST symbols.

Also perform these dlopen/dlsym intialization operations from
liblttng-ust-fd.so's contructor, so if close() is used from a signal
handler, it does not have to invoke dlopen()/dlsym() which are not
async-signal-safe.

Add compiler #error around closefrom() which is only emitted on Solaris
and BSDs, and not relevant to the 2.13 custom upgrade scope.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: I292434328f512bbb284c97f6a6170edbeef7c222